requirements.txt
python-dotenv
google-genai
fastapi
uvicorn
Pillow
firebase-admin
pydantic
passlib[bcrypt]